NF+RF: implement read and write support for BrainVoyager QX MSK files and refactor bv.py
